Module Client.Make

Parameters

module P : Mirage_clock.PCLOCK

module R : Resolver_mirage.S

module S : Conduit_mirage.S

Signature

module Connection : Cohttp_lwt.S.Connection

include Cohttp_lwt.S.Client with type ctx = Connection.Net.ctx

type ctx = Connection.Net.ctx

val set_cache : Cohttp_lwt.S.call -> unit

Provide a function used to process requests. Please see call. The provided function is only used when no ctx argument is passed to the convenience functions below.

val call : ?ctx:ctx -> ?headers:Http.Header.t -> ?body:Cohttp_lwt.Body.t -> ?chunked:bool -> Http.Method.t -> Uri.t -> (Http.Response.t * Cohttp_lwt.Body.t) Lwt.t

call ?ctx ?headers ?body ?chunked meth uri

returns
(response, response_body) Consume response_body in a timely fashion. Please see call about how and why.

parameter chunked
use chunked encoding if true. The default is false for compatibility reasons.

parameter ctx
If provided, no connection cache is used, but Connection_cache.Make_no_cache.create is used to resolve uri and create a dedicated connection with ctx.

In most cases you should use the more specific helper calls in the interface rather than invoke this function directly. See head, get and post for some examples.

val head : ?ctx:ctx -> ?headers:Http.Header.t -> Uri.t -> Http.Response.t Lwt.t

val get : ?ctx:ctx -> ?headers:Http.Header.t -> Uri.t -> (Http.Response.t * Cohttp_lwt.Body.t) Lwt.t

val delete : ?ctx:ctx -> ?body:Cohttp_lwt.Body.t -> ?chunked:bool -> ?headers:Http.Header.t -> Uri.t -> (Http.Response.t * Cohttp_lwt.Body.t) Lwt.t

val post : ?ctx:ctx -> ?body:Cohttp_lwt.Body.t -> ?chunked:bool -> ?headers:Http.Header.t -> Uri.t -> (Http.Response.t * Cohttp_lwt.Body.t) Lwt.t

val put : ?ctx:ctx -> ?body:Cohttp_lwt.Body.t -> ?chunked:bool -> ?headers:Http.Header.t -> Uri.t -> (Http.Response.t * Cohttp_lwt.Body.t) Lwt.t

val patch : ?ctx:ctx -> ?body:Cohttp_lwt.Body.t -> ?chunked:bool -> ?headers:Http.Header.t -> Uri.t -> (Http.Response.t * Cohttp_lwt.Body.t) Lwt.t

val post_form : ?ctx:ctx -> ?headers:Http.Header.t -> params:(string * string list) list -> Uri.t -> (Http.Response.t * Cohttp_lwt.Body.t) Lwt.t

val callv : ?ctx:ctx -> Uri.t -> (Http.Request.t * Cohttp_lwt.Body.t) Lwt_stream.t -> (Http.Response.t * Cohttp_lwt.Body.t) Lwt_stream.t Lwt.t

deprecated
use moduleCohttp_lwt.Connection instead.

val ctx : ?authenticator:X509.Authenticator.t -> R.t -> S.t -> ctx

val with_authenticator : X509.Authenticator.t -> ctx -> ctx